scripts/requirements.txt
openai>=1.2.0
requests>=2.28.0
feedparser>=6.0.10
python-dotenv>=1.0.0